requirements.txt
# Core
python-telegram-bot==20.7
httpx~=0.25.2
requests==2.31.0

# Crypto (GetContact encryption)
pycryptodome==3.20.0

# Data files
pandas==2.1.4
openpyxl==3.1.2

# Web scraping (dog owner search)
undetected-chromedriver==3.5.5
selenium==4.18.1
beautifulsoup4==4.12.3

# Image analysis (optional — comment out if not needed)
Pillow==10.2.0
aiohttp==3.9.3
torch==2.2.0
transformers==4.37.2